International bilateral fisheries treaties (1960-1970)

An abbreviated version of the international fisheries agreements dataset, including only bilateral agreements, fewer variables, and ranging only between 1960 and 1970 inclusive. This data set is only meant for testing, and not for inference. It provides an example of an undirected, weighted (by integer/increment) network, with composition change and both monadic and dyadic covariates. Monadic variables include the dates states gain or lose sovereign status, their polity score, and their GDP. Dyadic variables include bilateral fisheries agreements between states, and states' contiguity with one another over time.

data(Fisheries_Treaties_6070)

bilatchanges

bilatnet

contigchanges

contignet

gdpchanges

regchanges

sovchanges

states

Format

The data includes several dataframes: states (154 rows, 4 columns, monadic), sovchanges (62 rows, 3 columns, monadic), regchanges (145 rows, 3 columns, monadic), gdpchanges (979 rows, 3 columns, monadic), bilatchanges (77 rows, 4 columns, dyadic), contigchanges (139 rows, 4 columns, dyadic). See below for variables and formats.

Object	Description	Format

states$label	Node identifier labels	character

states$present	Node present in dataset	boolean

states$regime	Placeholder for regime variable	numeric (NA)

states$gdp	Placeholder for GDP variable	numeric (NA)

sovchanges$time	Date of state sovereignty update	POSIXct

sovchanges$node	Node for state sovereignty update	integer

sovchanges$replace	State sovereignty update	boolean

regchanges$time	Date of regime update	POSIXct

regchanges$node	Node for regime update	integer

regchanges$replace	Regime update	integer (-10–10)

gdpchanges$time	Date of GDP update	POSIXct

gdpchanges$node	Node for GDP update	integer

gdpchanges$replace	GDP update	numeric

bilatchanges$time	Date of bilateral change	POSIXct

bilatchanges$sender	First bilateral change node	integer

bilatchanges$receiver	Second bilateral change node	integer

bilatchanges$increment	Create or dissolve tie	numeric (-1 or 1)

contigchanges$time	Date of contiguity change	POSIXct

contigchanges$sender	First contiguity change node	integer

contigchanges$receiver	Second contiguity change node	integer

contigchanges$replace	New contiguity value	numeric

An object of class data.frame with 77 rows and 4 columns.

An object of class matrix (inherits from array) with 154 rows and 154 columns.

An object of class data.frame with 139 rows and 4 columns.

An object of class matrix (inherits from array) with 154 rows and 154 columns.

An object of class data.frame with 979 rows and 3 columns.

An object of class data.frame with 145 rows and 3 columns.

An object of class data.frame with 62 rows and 3 columns.

An object of class data.frame with 154 rows and 4 columns.

References

Hollway, James, and Johan Koskinen. 2016. Multilevel Embeddedness: The Case of the Global Fisheries Governance Complex. Social Networks, 44: 281-94. doi:10.1016/j.socnet.2015.03.001 .

Hollway, James, and Johan H Koskinen. 2016. Multilevel Bilateralism and Multilateralism: States' Bilateral and Multilateral Fisheries Treaties and Their Secretariats. In Multilevel Network Analysis for the Social Sciences, edited by Emmanuel Lazega and Tom A B Snijders, 315-32. Cham: Springer International Publishing. doi:10.1007/978-3-319-24520-1_13 .