RFID Validity dataset

Dataset collected at ETH Zürich by Timon Elmer and colleagues in order to test the accuracy of Radio Frequency Identification (RFID) badges for measuring social interactions. Social interactions of 11 individuals (from the university staff) were recorded with RFID badges in an informal setting. They were then compared to the interactions observed by two confederates who watched the video recording of the event. The RFID data went through the data processing procedure detailed in the original article. See Elmer et al, 2019 for more details, and the OSF platform for all details on the dataset.

data(RFID_Validity_Study)

rfid

video

known.before

participants

Format

3 dataframes:

participants (11 rows, 7 columns): attributes of the experiment's participants

rfid (1011 rows, 4 columns): dyadic interactions detected by the RFID badges (after data processing)

video (219 rows, 4 columns): dyadic interactions detected by the video rating
and one network:

known.before (11 rows, 11 columns): network of previous acquaintances
See below for variables and formats.

Object	Description	Format

participants$actor	Identifier of the actor	integer

participants$label	(Anonymized) name	Factor

participants$present	Presence of the actor (all actors are present)	logical

participants$age	Actor's age	integer

participants$gender	Actor's gender (0: male, 1: female)	integer

participants$group	Actor's group affiliation (groups have distinct ids)	integer

participants$level	Actor's seniority (1: MSc student, 2: PhD student, 3: PostDoc, 4: Prof)	integer

rfid$NodeA	Identifier for the first actor	chr

rfid$NodeB	Identifier for the second actor	chr

rfid$Start	Time of the beginning of the dyadic interaction	integer

rfid$End	Time of the end of the dyadic interaction	integer

video$NodeA	Identifier for the first actor	chr

video$NodeB	Identifier for the second actor	chr

video$Start	Time of the beginning of the dyadic interaction	integer

video$End	Time of the end of the dyadic interaction	integer

An object of class data.frame with 1011 rows and 4 columns.

An object of class data.frame with 219 rows and 4 columns.

An object of class matrix (inherits from array) with 11 rows and 11 columns.

An object of class data.frame with 11 rows and 7 columns.

Source

https://osf.io/rrhxe/

References

Elmer, T., Chaitanya, K., Purwar, P., & Stadtfeld, C. (2019). The validity of RFID badges measuring face-to-face interactions. Behavior research methods, 1-19. doi:10.3758/s13428-018-1180-y