Define dependent events for a model

The final step in defining the data objects is to identify the dependent events.

defineDependentEvents(events, nodes, nodes2 = NULL, defaultNetwork = NULL)

Arguments

events

a data frame containing the event list that should be considered as a dependent variable in models.

nodes

a data frame or a nodes.goldfish object containing the nodes used in the event list

nodes2

a second nodeset in the case of two mode events

defaultNetwork

the name of a goldfish network object

Value

an object of class dependent.goldfish

Details

Before this step is performed, we have to define: 1. the nodeset (defineNodes), the network (defineNetwork) and the eventlist of the network (linkEvents).

See also

Examples

actors <- data.frame(
  actor = 1:5, label = paste("Actor", 1:5),
  present = TRUE, gender = sample.int(2, 5, replace = TRUE)
)
actors <- defineNodes(nodes = actors)
calls <- data.frame(
  time = c(12, 27, 45, 56, 66, 68, 87), sender = paste("Actor", c(1, 3, 5, 2, 3, 4, 2)),
  receiver = paste("Actor", c(4, 2, 3, 5, 1, 2, 5)), increment = rep(1, 7)
)
callNetwork <- defineNetwork(nodes = actors)
callNetwork <- linkEvents(x = callNetwork, changeEvent = calls, nodes = actors)

# Defining the dependent events:
callDependent <- defineDependentEvents(events = calls, nodes = actors, defaultNetwork = callNetwork)