Calculation for estimating an DyNAM-rate model

Given input data, it output the derivative of the loglikelihood, the Fisher information matrix, the logLikelihood, and the loglikelihood of each event

estimate_DyNAM_rate(
  parameters,
  dep_event_mat,
  timespan,
  is_dependent,
  stat_mat_init,
  stat_mat_update,
  stat_mat_update_pointer,
  stat_mat_rightcensored_update,
  stat_mat_rightcensored_update_pointer,
  presence1_init,
  presence1_update,
  presence1_update_pointer,
  presence2_init,
  presence2_update,
  presence2_update_pointer,
  n_actors_1,
  n_actors_2,
  twomode_or_reflexive,
  impute = TRUE
)

Arguments

parameters

An n_parameters by 1 matrix, which is the input parameter

dep_event_mat

An 2 by n_events matrix, the (1,n) entry is the sender of n-th event, and the (2,n) entry is the receiver of the n-th event.

timespan

An n_events by 1 matrix. The i-th element is the waiting time of the i-th event.

is_dependent

An n_events by 1 matrix with boolean values. If the i-th event is dependent, then the i-th entry of is_dependent is TRUE, otherwise it's FALSE.

stat_mat_init

An n_actor1*n_actor2 by n_parameters matrix. It is the initialization of the statistics matrix. The initial value of k-th effect for the actor1-actor2 pair (i,j) is recorded in the ((i-1)*n_actor2 + j,k) entry of stat_mat_init.

stat_mat_update

An matrix with four rows, which records the updates of the statistics matrix through all dependent events. The following is an example.

0	2	0	9	9

2	1	5	4	3

3	5	5	0	5

1.2	3.5	2.5	9.23	2.8

Each column represents an update. For example the first column means replacing the value of the (3+1)-th effect for the actor1-actor2 pair(0+1,2+1) by 1.2. The +1 is due to the difference between the numberings in R and C.

stat_mat_update_pointer

An n_events by 1 matrix that record which update belongs to which rightcensored event. The structure is similar to stat_mat_update_pointer.

stat_mat_rightcensored_update

An matrix with four rows, which record the updates of the statistics matrix through all right-censored events. the structure is similar to stat_mat_update.

presence1_init

An n_actor1 by 1 matrix, which records the initial presence of each actor1. If the i-th actor1 is not present in the beginning then the i-th entry of presence1_init is 0, otherwise it's 1.

presence1_update

An matrix with two rows, which record the updates of the presence of actor1 through all events. The following is an example.

0	3	4	9	20

1	0	1	0	0

Each column represents an update. For example the first column means the 0+1-th actor1 becomes present. And the second column means the the 3+1 th actor1 becomes absent. The +1 is due to the difference between the numberings in R and C.

presence2_init

An n_actors2 by 1 matrix, which records the initial presence of each actor2. If the i-th actor2 is not present in the beginning then the i-th entry of presence2_init is 0, otherwise it's 1.

presence2_update

An matrix with two rows, which record the updates of the presence of actor2 through all events. The following is an example.

0	3	4	9	20

1	0	1	0	0

Each column represents an update. For example the first column means the 0+1-th actor2 becomes present. And the second column means the the 3+1 th actor2 becomes absent. The +1 is due to the difference between the numberings in R and C.

presence2_update_pointer

An n_events by 1 matrix that record which update belongs to which (dependent+ rightcensored) event. The structure is similar to stat_mat_update_pointer.

n_actors_1

An integer which is the number of actor1

n_actors_2

An integer which is the number of actor2

twomode_or_reflexive

An boolean variable. If it's true, then the model is two-model or we consider the reflexive effect (that's the value in the diagonal entries of the statistics matrix).

impute

An boolean variable. If it true, the function does the imputation for missing values.

Value

Return a list with elements as follows.

derivative

An 1 by n_parameters matrix, which is the derivative of loglikelihood given the input parameter and data.

fisher

An n_parameters by n_parameters matrix, which is the fisher information matrix given the input parameter and data.

logLikelihood

An scalar, which is the loglikelihood given the input parameter and data.

intervalLogL

An n_events by 1 matrix, of which the i-th entry is the loglikelihood of the i-th event given the input parameter and data.