Gathering data for sender receiver model

Gathering data for model that for choosing an receiver given a sender, i.e. DyNAM-choice model. Only the useful information are recorded: Consider a actor1-actor2 pair, if the actor1 is not the sender of the current event or the actor2 is not present, then the information is not collected.

gather_receiver_model(
  dep_event_mat,
  stat_mat_init,
  stat_mat_update,
  stat_mat_update_pointer,
  presence2_init,
  presence2_update,
  presence2_update_pointer,
  n_actor1,
  n_actor2,
  twomode_or_reflexive,
  verbose = FALSE,
  impute = TRUE
)

Arguments

dep_event_mat

An 2 by n_events matrix, the (1,n) entry is the sender of n-th event, and the (2,n) entry is the receiver of the n-th event.

stat_mat_init

An n_actor1*n_actor2 by n_parameters matrix. It is the initialization of the statistics matrix. The initial value of k-th effect for the actor1-actor2 pair (i,j) is recorded in the ((i-1)*n_actor2 + j,k) entry of stat_mat_init.

stat_mat_update

An matrix with four rows, which records the updates of the statistics matrix through all dependent events. The following is an example.

0	2	0	9	9

2	1	5	4	3

3	5	5	0	5

1.2	3.5	2.5	9.23	2.8

Each column represents an update. For example the first column means replacing the value of the (3+1)-th effect for the actor1-actor2 pair(0+1,2+1) by 1.2. The +1 is due to the difference between the numberings in R and C.

stat_mat_update_pointer

An n_events by 1 matrix that record which update belongs to which rightcensored event. The structure is similar to stat_mat_update_pointer.

presence2_init

An n_actors2 by 1 matrix, which records the initial presence of each actor2. If the i-th actor2 is not present in the beginning then the i-th entry of presence2_init is 0, otherwise it's 1.

presence2_update

An matrix with two rows, which record the updates of the presence of actor2 through all events. The following is an example.

0	3	4	9	20

1	0	1	0	0

Each column represents an update. For example the first column means the 0+1-th actor2 becomes present. And the second column means the the 3+1 th actor2 becomes absent. The +1 is due to the difference between the numberings in R and C.

presence2_update_pointer

An n_events by 1 matrix that record which update belongs to which (dependent+ rightcensored) event. The structure is similar to stat_mat_update_pointer.

twomode_or_reflexive

An boolean variable. If it's true, then the model is two-model or we consider the reflexive effect (that's the value in the diagonal entries of the statistics matrix).

verbose

An boolean variable. It it's true, the function print the progress to the screen.

impute

An boolean variable. If it true, the function does the imputation for missing values.

Value

Return a list with elements as follows. The meaning of the argument can be found in corresponding computation codes, e.g. compute_coordination_selection.cpp.

stat_all_events

n_candidates

n_candidates1

n_candidates2

selected

selected_actor1

selected_actor2