goldfish package

The goldfish Project is an R package that allows to fit statistical network models (such as DyNAM and REM) to dynamic network data.

Details

The goldfish package in R allows the study of time-stamped network data using a variety of models. In particular, it implements different types of Dynamic Network Actor Models (DyNAMs), a class of models that is tailored to the study of actor-oriented network processess through time. Goldfish also implements different versions of the tie-oriented Relational Event Model by Carter Butts.

References

Stadtfeld, C. (2012). Events in Social Networks: A Stochastic Actor-oriented Framework for Dynamic Event Processes in Social Networks. KIT Scientific Publishing. doi:10.5445/KSP/1000025407

Stadtfeld, C., and Block, P. (2017). Interactions, Actors, and Time: Dynamic Network Actor Models for Relational Events. Sociological Science 4 (1), 318-52. doi:10.15195/v4.a14

Stadtfeld, C., Hollway, J., and Block, P. (2017). Dynamic Network Actor Models: Investigating Coordination Ties Through Time. Sociological Methodology 47 (1). doi:10.1177/0081175017709295

Hollway, J. (2020). Network embeddedness and the rate of international water cooperation and conflict. In Networks in Water Governance, edited by Manuel Fischer and Karin Ingold. London: Palgrave, pp. 87-113.

Hoffman, M., Block P., Elmer T., and Stadtfeld C. (2020). A model for the dynamics of face-to-face interactions in social groups. Network Science, 8(S1), S4-S25. doi:10.1017/nws.2020.3

See also

Author

Maintainer: Alvaro Uzaheta ualvaro@ethz.ch (ORCID)

Authors:

James Hollway james.hollway@graduateinstitute.ch (ORCID) (IHEID) [data contributor]

Christoph Stadtfeld c.stadtfeld@ethz.ch [data contributor]

Marion Hoffman

Other contributors:

Mirko Reul [contributor]

Timon Elmer [contributor]

Kieran Mepham [contributor]

Per Block [contributor]

Xiaolei Zhang [contributor]

Weigutian Ou [contributor]

Emily Garvin [contributor]

Siwei Zhang [contributor]

Mabel Wylie [contributor]