Goldfish effects.

This page catalogues the list of generic effects that are available in the goldfish package. ignoreRep, weighted, window, and transformed versions (transformFun) can be included by adding these arguments.

# structural effects: dyads statistics
tie(network, weighted = FALSE, window = Inf, ignoreRep = FALSE, transformFun = identity)
inertia(network, weighted = FALSE, window = Inf, transformFun = identity)
recip(network, weighted = FALSE, window = Inf, ignoreRep = FALSE, transformFun = identity)
trans(network, window = Inf, ignoreRep = FALSE, transformFun = identity)
mixedTrans(list(network1, network2), window = Inf, ignoreRep = FALSE, transformFun = identity)
four(network, isTwoMode = FALSE, window = Inf, ignoreRep = FALSE, transformFun = identity)
# structural effects: node statistics
indeg(network, isTwoMode = FALSE, weighted = FALSE, window = Inf, ignoreRep = FALSE,
      type = c("alter", "ego"), transformFun = identity)
outdeg(network, isTwoMode = FALSE, weighted = FALSE, window = Inf, ignoreRep = FALSE,
      type = c("alter", "ego"), transformFun = identity)
node_trans(network, isTwoMode = FALSE, window = Inf, ignoreRep = FALSE,
      type = c("alter", "ego"), transformFun = identity)
# attribute effects
ego(attribute)
alter(attribute)
same(attribute)
diff(attribute, transformFun = abs)
sim(attribute, transformFun = abs)
# structural + attribute effects
tertius(network, attribute, isTwoMode = FALSE, weighted = FALSE, window = Inf,
        ignoreRep = FALSE, type = c("alter", "ego"), transformFun = identity,
        aggregateFun = function(x) mean(x, na.rm = TRUE))
tertius_diff(network, attribute, isTwoMode = FALSE, weighted = FALSE, window = Inf,
             ignoreRep = FALSE, transformFun = abs,
             aggregateFun = function(x) mean(x, na.rm = TRUE))

Arguments

network, attribute
Objects upon which the effect should be calculated. For network effects, this would be either an independent network or the dependent network by default. For attribute effects, this should be the attribute name with the indication of the data frame containing the initial values of nodes attributes (data.frame$attribute), see defineNodes for details.

ignoreRep
Identifies whether the effect recognizes actors to send additional ties beyond the first to receivers (FALSE means additional ties are taken into account). The default value is FALSE.

weighted
Identifies whether the effect relies on the presence or number of ties (TRUE means it relies also on the number of ties). The interpretation depends on the specific effect. The default value is FALSE.

window
Identifies a window length within which changes should apply to events. The default value is Inf meaning that no windows are applied to the effect. Window size can be specified as a number in seconds (i.e. an hour is 3600 seconds), or as a string stating the number of time units in the format "number unit", for example "30 seconds", "1 minute", "2 hours", "3 days", "4 weeks", "5 months", or "6 years".

transformFun
Use this parameter to obtain transformed statistics. The argument expected is a function or the name of the function to use without quotation marks.

aggregateFun
Use this parameter to obtain an aggregated statistics for indirect effects like tertius and tertius_diff. The argument expected is a function or the name of the function to use without quotation marks.

type
Applies only to indegree, outdegree and node_trans in the case of the REM model. In the default case of type = "alter", the effect returns statistics matrices according to the potential receivers. In the case where type = "ego", the effect returns statistics matrices according to the sender. This argument does not apply in the case of the DyNAM models: the DyNAM-choice model only considers the "alter" value, while in the DyNAM-rate, the "ego" value is always to be considered for these three effects.

Value

List with the changes to a statistics matrix. It also can have an additional component with the cache object if the effect computation requires one.

Details

Note that the use of some effects (combinations) are ill-advised. For example, using tie(network, ignoreRep = FALSE), where the network refers to the dependent network, will always result in a change statistic of zero, and thus cannot be used.

Functions

Structural effects (dyads statistics): These effects cannot be used for the DyNAM-rate model (model = "DyNAM" and subModel = "rate").

tie

Tendency to create an event i->j if the tie i->j exists in a covariate network. Parameter weighted can be set to TRUE if the weight of j->i is to be taken as a statistic. It can be transformed by using transformFun (This only makes sense with weighted = TRUE).

inertia

Usually used as the "intercept" for the choice function, 'inertia' is the tendency to create an event i->j if the event i->j happened before. It can be interpreted as the differential tendency to update existing ties rather than creating new ones. Thus, 'inertia' is similar to 'tie', but defined on the network to which the dependent events relate. Parameter weighted can be set to TRUE if the weight of j->i is to be taken as a statistic. It can be transformed by using transformFun (This only makes sense with weighted = TRUE). Note: inertia can never be used in combination with a ignoreRep = TRUE parameter as this would replace all positive statistics with zeros.

recip

Reciprocity. Tendency to create an event i->j if one or several j->i happened before. Parameter weighted can be set to TRUE if the weight of j->i is to be taken as a statistic. It can be transformed by using transformFun (This only makes sense with weighted = TRUE). This effect cannot be used for two-mode networks.

trans

Transitivity. It is the tendency to create an event i->j if it closes a two-paths (i->k->j) in the past events network. It can be transformed by using transformFun. This effect cannot be used for two-mode networks.

mixedTrans

Transitivity within 2 networks. It is the tendency to create an event i->j if it closes a two-paths with events (i->k) in network1 and (k->j) in network2 in the past events. It can be transformed by using transformFun. This effect cannot be used for two-mode networks.

four

It describes the tendency to create an event i->j if there are many possible closures of three-paths (i->k<-l->j) in the past events network.

Structural effects (node statistics):

indeg

In the DyNAM-Rate model: tendency of actor i to send a tie when i has a high incoming degree in a covariate network (ego type). In the DyNAM-choice model: tendency to create an event i->j when j has a high incoming degree in a covariate network ('alter' type). In the REM model: tendency to create an event i->j when either i or j has a high incoming degree in a covariate network. An argument type allows to choose whether we want to use the indegree effect for sender i (type = "ego") or for receiver j (type = "alter"). The degree can be transform with transformFun.

outdeg

In the DyNAM-Rate model: tendency of actor i to send a tie when i has a high outgoing degree in a covariate network (ego type). In the DyNAM-choice model: tendency to create an event i->j when j has a high outgoing degree in a covariate network ('alter' type). In the REM model: tendency to create an event i->j when either i or j has a high outgoing degree in a covariate network. An argument type allows to choose whether we want to use the outdegree effect for sender i (type = "ego") or for receiver j (type = "alter"). The degree can be transform with transformFun.

node_trans

Embeddedness in transitive structures as a source node. It is the tendency of actor i to send a tie when there are transitive triangles where i is the source (i->k->j<-i) that are closed. The degree can be transform with transformFun.

Attribute effects:

ego

In the DyNAM-rate is the tendency of actors to be more active when they score high on attribute. For REM is the tendency of an event i->j when i has a high attribute. This effects cannot be used for the DyNAM-choice model (model = "DyNAM" and subModel = "choice")

alter

The tendency to create an event i->j when j has a high attribute. These effects cannot be used for the DyNAM-rate model (model = "DyNAM" and subModel = "rate")

same

Homophily (same value). The tendency of an event i->j to happen if actors i and j have the same attribute. This effect cannot be used for two-mode networks and for the DyNAM-rate model (model = "DyNAM" and subModel = "rate")

sim

Homophily (similar value). The tendency to create an event i->j where both actors have similar attribute values (low absolute differences regarding attribute if transformFun = abs). This effect cannot be used for two-mode networks and neither for the DyNAM-rate model (model = "DyNAM" and subModel = "rate")

diff

Heterophily. The tendency to create an event i->j where both actors have different attribute values (high absolute differences regarding attribute if transformFun = abs). This effect cannot be used for two-mode networks and neither for the DyNAM-rate model (model = "DyNAM" and subModel = "rate")

Structural + attribute effects (node or dyads statistics):

tertius

For DyNAM-choice and REM is the tendency to create an event i->j when j has a high aggregate (aggregateFun) value of its in-neighbors (all k with network[k, j] > 0). For DyNAM-rate is the tendency of actors to be more active when they have a high aggregate value of their in-neighbors. Note: When the node does not have in-neighbors, the average of the aggregate values of nodes is used for the node.

tertius_diff

The tendency to create an event i->j when i has a similar value as j aggregate (aggregateFun) value of its in-neighbors (all k with network[k, j] > 0. Note: When the node j does not have in-neighbors, the average of the similarities computed for the pairs i, k when k has in-neighbors is used for the node.

Examples

data("Social_Evolution")
actors<-defineNodes(nodes = actors)
call.Network <- defineNetwork(nodes = actors, directed = TRUE)
call.Network <- linkEvents(x = call.Network, changeEvent = calls, nodes = actors)
calls<-defineDependentEvents(events=calls, nodes=actors,
                             defaultNetwork = call.Network)

# Using a DyNAM-Rate model:
estimate(calls ~ ego(actors$gradeType) + indeg(call.Network),
         model = "DyNAM", subModel = "rate")

#> Parsing formula.

#> Error: The left hand side of the formula should contain dependent events (check the function defineDependentEvents).

# Using a DyNAM-Choice model:
estimate(calls ~ trans(call.Network, window=259200) + indeg(call.Network),
         model = "DyNAM", subModel = "choice")

#> Parsing formula.

#> Error: The left hand side of the formula should contain dependent events (check the function defineDependentEvents).

# Using a REM model:
estimate(calls ~ outdeg(call.Network, type="ego") +
                indeg(call.Network, type="alter"), model = "REM")

#> Parsing formula.

#> Error: The left hand side of the formula should contain dependent events (check the function defineDependentEvents).