Printing functions for `goldfish` objects.

Printing functions for `goldfish` objects.

# S3 method for result.goldfish
print(
  x,
  ...,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width")
)

# S3 method for result.goldfish
summary(object, ...)

# S3 method for summary.result.goldfish
print(
  x,
  ...,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  fixed = FALSE
)

# S3 method for nodes.goldfish
print(x, ..., full = FALSE, n = 6)

# S3 method for nodes.goldfish
head(x, ..., n = 6L)

# S3 method for nodes.goldfish
tail(x, ..., n = 6L, keepnums = FALSE, addrownums = FALSE)

# S3 method for network.goldfish
print(x, ..., full = FALSE, n = 6L)

# S3 method for network.goldfish
head(x, ..., n = 6L)

# S3 method for network.goldfish
tail(x, ..., n = 6L, keepnums = TRUE, addrownums = TRUE)

# S3 method for dependent.goldfish
print(x, ..., full = FALSE, n = 6)

# S3 method for dependent.goldfish
head(x, ..., n = 6L)

# S3 method for preprocessed.goldfish
print(x, ..., width = getOption("width"))

Arguments

x

an object of class result.goldfish, summary.result.goldfish, nodes.goldfish, network.goldfish, dependent.goldfish, or preprocessed.goldfish.

...

further arguments to be passed to the respective default method.

digits

minimal number of significant digits, see print.default.

width

only used when max.levels is NULL, see print.

object

an object of class result.goldfish.

fixed

logical. Indicates whether the parameter coefficients of effects fixed during estimation using `fixedParameters` should be printed. The default value is TRUE. Note: applies for objects of class result.goldfish and summary.result.goldfish.

full

logical. Indicates whether the complete matrix/data.frame should be printed. The default value FALSE.

n

number of rows for data.frame, and rows and columns for matrix to be printed.

keepnums

in each dimension, if no names in that dimension are present, create them using the indices included in that dimension. Ignored if dim(x) is NULL or its length 1.

addrownums

deprecated - keepnums should be used instead. Taken as the value of keepnums if it is explicitly set when keepnums is not.