Social evolution of a university dormitory cohort

An abbreviated version of the MIT Reality Commons Social Evolution dataset, spanning a reduced time period and with fewer variables. Dyadic variables include binary friendships at time of survey, and time-stamped phone call occurrences. Individual variables include the floor of the dormitory on which the student resides, and the grade type of each student including freshmen, sophomore, junior, senior, or graduate tutors.

data(Social_Evolution)

Format

3 dataframes: actors (84 rows, 4 columns), calls (439 rows, 4 columns), friendship (766 rows, 4 columns). See below for variables and formats.

Object	Description	Format

actors$label	Actor identifier labels	character

actors$present	Actor present in dataset	boolean

actors$floor	Floor of residence actor lives on	numeric (1-9)

actors$gradeType	Degree level	numeric (1-5)

calls$time	Time and date of call	numeric from POSIXct

calls$sender	Initiator of phone call	character

calls$receiver	Recipient of phone call	character

calls$increment	Indicates call number increment (all 1s)	numeric (1)

friendship$time	Time and date of friend nomination	numeric from POSIXct

friendship$sender	Nominator of friendship	character

friendship$receiver	Nominee of friendship	character

friendship$replace	Indicates friendship value at $time	numeric

Source

http://realitycommons.media.mit.edu/socialevolution.html

References

A. Madan, M. Cebrian, S. Moturu, K. Farrahi, A. Pentland (2012). Sensing the 'Health State' of a Community. Pervasive Computing. 11, 4, pp. 36-45.