Estimate a model

Estimates parameters for a dynamic network model via maximum likelihood.

estimate(x, model = c("DyNAM", "REM"),
         subModel = c("choice", "rate", "choice_coordination"),
         estimationInit = NULL, preprocessingInit = NULL,
         preprocessingOnly = FALSE,
         verbose = FALSE, silent = FALSE, debug = FALSE)

Arguments

x
a formula that defines at the left-hand side the dependent network (see defineDependentEvents) and at the right-hand side the effects and the variables for which the effects are expected to occur (see goldfishEffects).

model
a character string defining the model type. Current options include "DyNAM" or "REM"

DyNAM

Dynamic Network Actor Models (Stadtfeld, Hollway and Block, 2017 and Stadtfeld and Block, 2017)

REM

Relational Event Model (Butts, 2008)

subModel
a character string defining the submodel type. Current options include "choice", "rate" or "choice_coordination"

choice

a multinomial receiver choice model if model = "DyNAM" (Stadtfeld and Block, 2017), or the general Relational event model if model = "REM" (Butts, 2008).

choice_coordination

a multinomial-multinomial model for coordination ties (Stadtfeld, Hollway and Block, 2017)

rate

A individual activity rates model if model = "DyNAM" (Stadtfeld and Block, 2017)

estimationInit
a list containing lower level technical parameters for estimation. May be:

initialParameters

a numeric vector. It includes initial parameters of the estimation. Default is set to NULL.

fixedParameters

a numeric vector. It specify which component of the parameter (intercept included) is fixed and what's the fixed values during the estimation. E.g. if the vector is c(2,NA). Then the first component of the parameter is fixed to 2 during the whole esitmation process. Default is set to NULL, i.e. we don't fixed any parameter. Note that it must be consistent with initialParameters.

maxIterations

maximum number of iterations of the Gauss/Fisher scoring method for the estimation. Default is set to 20.

maxScoreStopCriterion

maximum absolute score criteria for successful convergence. Default value is 0.001

initialDamping

a numeric vector used to declare the initial damping factor for each parameter. It controls the size of the update step during the iterative estimation process. The default is set to 30 when the formula has windowed effects or 10 in another case, see goldfishEffects.

dampingIncreaseFactor

a numeric value. It controls the factor that increases the damping of the parameters when improvements in the estimation are found.

dampingDecreaseFactor

a numeric value. Controls the factor that decreases the damping of the parameters when no improvements in the estimation are found.

engine

a string indicating the estimation engine to be used. Current options include "old", "default", "default_c",and "gather_compute". The default value is "default", it is an estimation routine implemented in pure R code. "old" uses an R estimation routine that uses an inefficient version of data structures. "default_c" uses a C implementation of the "default" routine. "gather_compute" uses a C implementation with a different data structure that reduces the time but it can increase the memory usage.

startTime

a numerical value or a date-time character with the same time-zone formatting as the times in event that indicates the starting time to be considered during estimation.

endTime

a numerical value or a date-time character with the same time-zone formatting as the times in event that indicates the end time to be considered during estimation.

preprocessingInit
a preprocessed.goldfish object computed for the current formula, allows skipping the preprocessing step.

preprocessingOnly
logical. Indicates whether only preprocessed statistics should be returned rather than a results object.

verbose
logical indicating whether details of the estimation routine should be provided.

silent
logical indicating whether a mimimal output should be given.

debug
logical indicating whether very detailed intermediate results should be given; slows down the routine significantly.

Value

returns an object of class "result.goldfish" when preprocessingOnly = FALSE or a preprocessed statistics object of class "preprocessed.goldfish" when preprocessingOnly = TRUE.

An object of class "result.goldfish" is a list including:

parameters

a matrix with the coefficients estimates.

standardErrors

a vector with the standard errors of the coefficients.

logLikelihood

the log likelihood of the estimate model

finalScore

a vector with the final score reach by the parameters during estimation.

finalInformationMatrix

a matrix with the final values of the negative Fisher information matrix. The inverse of this matrix gives the variance-covariance matrix for the parameters estimates.

convergence

a list with two elements. The first element (isConverged) is a logical value that indicates the convergence of the model. The second element (maxAbsScore) reports the final maximum absolute score in the final iteration.

nIterations

an integer with the total number of iterations performed during the estimation process.

nEvents

an integer reporting the number of events considered in the model.

names

a matrix with a description of the effects used for model fitting. It includes the name of the object used to calculate the effects and additional parameter description.

formula

a formula with the information of the model fitted.

model

a character vector of the model type.

rightCensored

a logical value indicating if the estimation process considered right censored events. Only it is considered when model = "DyNAM", subModel = "rate" and the model includes intercept.

Details

Missing data is imputed during the preprocessing stage. For network data missing values are replaced by a zero value, it means that is assuming a not tie/event explicitly. For attributes missing values are replaced by the mean value, if missing values are presented during events updates they are replace by the mean of the attribute in that moment of time.

DyNAM

The actor-oriented models that the goldfish package implements have been called Dynamic Network Actor Models (DyNAMs). The model is a two-step process. In the first step, the waiting time until an actor \(i\) becomes active to send an event is modeled (model = "DyNAM" and subModel = "rate"). In the second step, the conditional probability of \(i\) choosing \(j\) as the event receiver is modeled (model = "DyNAM" and subModel = "choice"). These two-steps are assumed to be conditionally independent given the process state (Stadtfeld, 2012), due to this assumption is possible to estimate these components by different calls of the estimate function.

When model = "DyNAM" and subModel = "rate" is used to estimate the first step component of the process. It is important to add a time intercept to model the waiting times between events, in this way the model considers the right-censored intervals in the estimation process.

References

Butts C. (2008). A Relational Event Framework for Social Action. Sociological Methodology 38 (1). doi: 10.1111/j.1467-9531.2008.00203.x

Stadtfeld, C. (2012). Events in Social Networks: A Stochastic Actor-oriented Framework for Dynamic Event Processes in Social Networks. KIT Scientific Publishing. doi: 10.5445/KSP/1000025407

Stadtfeld, C., and Block, P. (2017). Interactions, Actors, and Time: Dynamic Network Actor Models for Relational Events. Sociological Science 4 (1), 318-52. doi: 10.15195/v4.a14

Stadtfeld, C., Hollway, J., and Block, P. (2017). Dynamic Network Actor Models: Investigating Coordination Ties Through Time. Sociological Methodology 47 (1). doi: 10.1177/0081175017709295

See also

Examples

# A multinomial receiver choice model
data("Social_Evolution")
callNetwork <- defineNetwork(nodes = actors, directed = TRUE)
callNetwork <- linkEvents(x = callNetwork, changeEvent = calls, nodes = actors)
callsDependent <- defineDependentEvents(events = calls, nodes = actors,
                                        defaultNetwork = callNetwork)
mod01 <- estimate(callsDependent ~ inertia + recip + trans,
                  model = "DyNAM", subModel = "choice")

#> Parsing formula.

#> Error in get(depName): object 'callsDependent' not found

summary(mod01)

#> Error in summary(mod01): object 'mod01' not found

# A individual activity rates model
mod02 <- estimate(callsDependent ~ 1 + node_trans + indeg + outdeg,
                  model = "DyNAM", subModel = "rate")

#> Parsing formula.

#> Error in get(depName): object 'callsDependent' not found

summary(mod02)

#> Error in summary(mod02): object 'mod02' not found

# A multinomial-multinomial choice model for coordination ties
data("Fisheries_Treaties_6070")
states <- defineNodes(states)
states <- linkEvents(states, sovchanges, attribute = "present")
states <- linkEvents(states, regchanges, attribute = "regime")
states <- linkEvents(states, gdpchanges, attribute = "gdp")

bilatnet <- defineNetwork(bilatnet, nodes = states, directed = FALSE)
bilatnet <- linkEvents(bilatnet, bilatchanges, nodes = states)

createBilat <- defineDependentEvents(
  events = bilatchanges[bilatchanges$increment == 1, ],
  nodes = states, defaultNetwork = bilatnet
)

contignet <- defineNetwork(contignet, nodes = states, directed = FALSE)
contignet <- linkEvents(contignet, contigchanges, nodes = states)
partner.model <- estimate(createBilat ~
  inertia(bilatnet) +
  indeg(bilatnet, ignoreRep = TRUE) + trans(bilatnet, ignoreRep = TRUE) +
  tie(contignet, ignoreRep = TRUE) +
  alter(states$regime, ignoreRep = TRUE) + diff(states$regime, ignoreRep = TRUE) +
  alter(states$gdp, ignoreRep = TRUE) + diff(states$gdp, ignoreRep = TRUE),
model = "DyNAM", subModel = "choice_coordination",
estimationInit = list(initialDamping = 40, maxIterations = 30)
)

#> Parsing formula.

#> Error in get(depName): object 'createBilat' not found

summary(partner.model)

#> Error in summary(partner.model): object 'partner.model' not found