Methods for goldfish objects.

Printing functions for goldfish objects.

# S3 method for result.goldfish
print(
  x,
  ...,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  complete = FALSE
)

# S3 method for summary.result.goldfish
print(
  x,
  ...,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  complete = FALSE
)

# S3 method for nodes.goldfish
print(x, ..., full = FALSE, n = 6)

# S3 method for network.goldfish
print(x, ..., full = FALSE, n = 6L)

# S3 method for dependent.goldfish
print(x, ..., full = FALSE, n = 6)

# S3 method for preprocessed.goldfish
print(x, ..., width = getOption("width"))

Arguments

x

an object of class result.goldfish, summary.result.goldfish, nodes.goldfish, network.goldfish, dependent.goldfish, or preprocessed.goldfish.

...

further arguments to be passed to the respective default method.

digits

minimal number of significant digits, see print.default().

width

controls the maximum number of columns on a line used in printing summary.result.goldfish and preprocessed.goldfish, see print.default().

complete

logical. Indicates whether the parameter coefficients of effects fixed during estimation using fixedParameters should be printed. The default value is FALSE. Note: applies for objects of class result.goldfish and summary.result.goldfish.

full

logical. Indicates whether the complete matrix/data.frame should be printed. The default value FALSE.

n

number of rows for data.frame, and rows and columns for matrix to be printed.

Value

Not value, called for printing side effect.

For objects of class result.goldfish and summary.result.goldfish

print the estimated coefficients when complete = FALSE, otherwise it includes also the fixed coefficients. For summary.result.goldfish print:

Effect details:

a table with additional information of the effects. The information corresponds to the values of the effects arguments when they are modified and if they where fixed during estimation, see vignette("goldfishEffects") for the complete list of arguments, and estimate() on how to fix coefficients during estimation.

Coefficients:

a table with the estimated coefficients, their approximate standard error obtain from the inverse of the negative Fisher information matrix, z-value and the p-value of the univariate two-tailed Wald test to test the hypothesis that the parameter is 0.

Convergence and Information Criteria:

Information about the convergence of the iterative Newton-Raphson procedure and the score value in the last iteration. Information criteria as the AIC, BIC and the AIC corrected for small sample size AICc are reported.

Model and subModel:

the values set during estimation.

For objects of class nodes.goldfish print information of the total number of nodes in the object, the number of nodes present at the beginning of preprocessing, a table with the linked attributes with their respective events data frame and a printing of the first rows in the nodes data frame. See defineNodes().

For objects of class network.goldfish print information of the dimensions of the network, number of ties presented at the beginning of the preprocessing, the nodes data frames linked to it, information about their definition as a one-mode and directed network, linked events data frame to it and a printing of the first rows and columns in the array. See defineNetwork().

For objects of class dependent.goldfish print information of the total number of events in the object, linked nodes set(s), linked default network and a printing of the first rows in the events data frame. See defineDependentEvents().